requirements.txt
PyQt6>=6.6.0
google-genai>=0.3.0
pydantic>=2.5.0
networkx>=3.2.0
qasync>=0.27.0
